onnx/requirements.txt
fastapi
numpy
pillow
onnxruntime
python-multipart